Refactor ASABConfig module to Maintenance
