+++ src/modules/maintenance/MicroservicesModule/MicroservicesContainers/MicroservicesContainer.js
@@ -31,7 +31,7 @@ export default (props) => {
 				}
 			}
 		},
-		{ name: t('MicroservicesContainer|ID'), key: 'id', link: { key: "id", pathname: "/config/svcs/" } },
+		{ name: t('MicroservicesContainer|ID'), key: 'id', link: { key: "id", pathname: "/microservices/svcs/" } },
 		{ name: t('MicroservicesContainer|Host'), key: 'hostname' },
 		{ name: t('MicroservicesContainer|Launch time'), key: 'launchtime', datetime: true },
 		{ name: t('MicroservicesContainer|Created at'), key: 'created_at', datetime: true },
